test(cojson): drop differential harness and shim test arms
Delete groupEngineDifferential.test.ts (its TS oracle is gone with the kill
switch). Rename shimNodeCore.test.ts to nodeCore.test.ts, dropping the
ShimNodeCore nodeCoreSuite arm and the shim dispose-semantics block; the
capability test now asserts the validation surface is present on every native
provider. Remove the bench's TS-fallback arms (native cases only).

Update the fixture exporter for the oracle inversion: with the kill switch
gone it captures native verdicts, so the committed fixtures are now frozen
golden files that regeneration must reproduce byte-for-byte in verdict content;
the seven previously-forced stubEnv calls are removed.
